# Don't call np.identity() in transforms.