@jklymak 2018-01-26
We are getting odd failures on 2.7 with some of your PRs now.  Not sure if it is "you" or "us"...

This change seems fine to me.